[Isolates] Merge revs 5516:5612 from bleeding-edge to Isolates branch. Review URL: http://codereview.chromium.org/3766005
git-svn-id: http://v8.googlecode.com/svn/branches/experimental@5626 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ isolates/src/heap-inl.h
@@ -45,7 +45,7 @@ void PromotionQueue::insert(HeapObject* target, int size) {
 
 
 void Heap::UpdateOldSpaceLimits() {
-  int old_gen_size = PromotedSpaceSize();
+  intptr_t old_gen_size = PromotedSpaceSize();
   old_gen_promotion_limit_ =
       old_gen_size + Max(kMinimumPromotionLimit, old_gen_size / 3);
   old_gen_allocation_limit_ =
